+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 This is the changelog of our "JTL WooCommerce Connector".
 
 ## Unreleased
-- CO-2461 - add button to clear connector cache
+- CO-2461 - add clear cache button, add disable cache button
+- fix debug log setting
 
 ## 1.40.3 _2023-10-25_
 - Fix Session is invalid error if session has been started before